collect fingers angle data
This works by using voltage divider rule, output were sent to any bluetooth serial terminal app that available in playstore. 
Change these value according to your specs.
R_DIV is the value of resistor that you use in the circuit.
flatResistance is the resistance value of flex sensor when it was left straight without any bending applied.
bendResistance is the resistance value of flex sensor when bending of 90 degree was applied to it.